Changed jogging record to reference API jogging data
When a NEW jogging record was created, the previous implementation of the createOrUpdateJogging() would use the jogging record  from the angular form, instead of using the jogging record being returned from the backend API. This resulted in jogging records being pushed to the JoggingData array that did NOT contain a valid Id, "undefined".  This change allows a user to edit and delete newly created jogging records without having to refresh the window to pull the Id from a get request.